ygot related enhancements and fixes (#34)
Enhanced the unmarshaling method to update the existing list entries instead of replacing it, and by providing this as one more option in the method.
Optimized the EmitJson method to improve the performance of the Marshaling of object into json data
Fixed the issue in finding the type, if the type is defined as typedf which got included as part of Include statement.
Added the support to synchronize the access of the pathToSchema, ChildSchema, regexp maps
Fixed the issue in MergeStruct method by comparing with src. object and updating the content of the dest. object properly
Added the lint:ignore statment to avoid static check on the methods getNodeName, getObjectFieldName
Fixed the issue to only skip the validation for sonic yang model, since the validation is handled in the CVL.